requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altgraph==0.17.5
attrs==26.1.0
evdev==1.9.3
frozenlist==1.8.0
idna==3.13
mss==10.2.0
multidict==6.7.1
packaging==26.2
pillow==12.2.0
propcache==0.4.1
pyinstaller==6.20.0
pyinstaller-hooks-contrib==2026.4
pynput==1.8.1
pytesseract==0.3.13
python-xlib==0.33
setuptools==82.0.1
six==1.17.0
yarl==1.23.0
